Add basic FujiHack menu - Add input map API to fujifilm.h - Add task API to fujifilm.h - Add a ton of stubs and numbers to XF1, for tasks and input
I was able to find the address for the input struct. It includes some
32 bit addresses related to key presses, the accelerometer, and the gryoscope.
Haven't dug into it too far but I researched enough to make a basic FujiHack menu.

I tried to figure out some of the multitasking related functions, but I didn't
get far at all. No matter what I tried, everything crashed the camera. This is
because of two possibilities:
- I'm not interfacing with the API correctly
- I can't create a task from a task.
I'm convinced it's the second.

Even if I could create a task, I would have to figure out how to
pause the UI task so that keypresses are only tracked by my custom
menu and not the UI.

I think the best temporary solution is to hijack the main "ui" task
and inject my code there. There I think I could properly overwrite on
top of the Fujifilm menus.

Also, this is my first time seeing how "bloaty" the camera is. It
tends to have glitchy menus and be sluggish in general. Not sure if
this is because the software is crap, or the CPU is crap.