Redesign build system I don't think I'll ever be content with any build system. This isn't too complicated, but It's hard to predict much it will grow in the future. I thought about using an alternative build tool like xmake, but decided against that since I don't want to learn a new tool.
This adds support for a minimal/ directory, which will be used for debugging
and research (instead of working in main src/). From now on, src/ will eventually
become the main binary, that might eventually be loaded in from file (?)

This also makes firm.c obsolete. It is kept on top level for documentation reasons,
mainly for the firmware header struct. Might eventually be put in a header file in src/