New stability fixes after testing huge software
For most of this project, I've only tested tiny 6kb binaries through USB.
Some new hacks allowed me to run Doom on my camera, which is around 10mb.

Doom port will be released on April first (for X-A2 only).

This is a huge step for testing the stability of this hack, and making sure
everything is POSIX compliant. More integration with the Frontier operating system
will follow.