GOT needs relocations even without PIC, fix linker bugs global data in the GCC libs need relocating. sprintf with %X didn't work before. Had a whole menu system working but git has one again fooled me and I lost my changes. It wasn't a whole lot so I should be able to rewrite it quickly.
Xa2 temporary code spaces is now 0x10000000, old area was a place the firmware used to load ROM.
I had thought it was unused after boot time, but it appears to be written once USB is unplugged.
New area is temporary until I can find an area that makes sense.