Add a new patch - load executable binary from menu This will allow huge files to be loaded into RAM and executed without having to plug in USB. This is meant to replace the INSTAX Settings menu option.
The file operations code will be run from the UIlib context, and it won't replace
any underlying functionality of the menu. This means the back button must be pressed
right after the INSTAX printer option in the menu.

file.S is ~53 instructions = 212 bytes, which is safe to inject into the 416 byte function
rendering the menu.